+++ platformio/device/monitor/command.py
@@ -13,14 +13,13 @@
 # limitations under the License.
 
 import os
-import sys
 
 import click
-from serial.tools import miniterm
 
 from platformio import exception, fs
 from platformio.device.finder import find_serial_port
 from platformio.device.monitor.filters.base import register_filters
+from platformio.device.monitor.terminal import start_terminal
 from platformio.platform.factory import PlatformFactory
 from platformio.project.config import ProjectConfig
 from platformio.project.exception import NotPlatformIOProjectError
@@ -30,10 +29,11 @@ from platformio.project.options import ProjectOptions
 @click.command("monitor", short_help="Monitor device (Serial/Socket)")
 @click.option("--port", "-p", help="Port, a number or a device name")
 @click.option(
-    "--baud",
     "-b",
+    "--baud",
     type=int,
-    help="Set baud rate, default=%d" % ProjectOptions["env.monitor_speed"].default,
+    default=ProjectOptions["env.monitor_speed"].default,
+    help="Set baud/speed, default=%d" % ProjectOptions["env.monitor_speed"].default,
 )
 @click.option(
     "--parity",
@@ -58,7 +58,9 @@ from platformio.project.options import ProjectOptions
     help="Set the encoding for the serial port (e.g. hexlify, "
     "Latin1, UTF-8), default: UTF-8",
 )
-@click.option("--filter", "-f", multiple=True, help="Add filters/text transformations")
+@click.option(
+    "-f", "--filter", "filters", multiple=True, help="Add filters/text transformations"
+)
 @click.option(
     "--eol",
     default="CRLF",
@@ -78,12 +80,20 @@ from platformio.project.options import ProjectOptions
     type=int,
     default=20,
     help="ASCII code of special character that is used to "
-    "control miniterm (menu), default=20 (DEC)",
+    "control terminal (menu), default=20 (DEC)",
 )
 @click.option(
     "--quiet",
     is_flag=True,
     help="Diagnostics: suppress non-error messages, default=Off",
+)
+@click.option(
+    "--reconnect/--no-reconnect",
+    default=True,
+    help=(
+        "If established connection fails, "
+        "silently retry on the same port, default=True"
+    ),
 )
 @click.option(
     "-d",
@@ -96,49 +106,32 @@ from platformio.project.options import ProjectOptions
     "--environment",
     help="Load configuration from `platformio.ini` and specified environment",
 )
-def device_monitor_cmd(**kwargs):  # pylint: disable=too-many-branches
-    project_options = {}
+def device_monitor_cmd(**options):
     platform = None
-    with fs.cd(kwargs["project_dir"]):
+    project_options = {}
+    with fs.cd(options["project_dir"]):
         try:
-            project_options = get_project_options(kwargs["environment"])
-            kwargs = apply_project_monitor_options(kwargs, project_options)
+            project_options = get_project_options(options["environment"])
+            options = apply_project_monitor_options(options, project_options)
             if "platform" in project_options:
                 platform = PlatformFactory.new(project_options["platform"])
         except NotPlatformIOProjectError:
             pass
-        register_filters(platform=platform, options=kwargs)
-        kwargs["port"] = find_serial_port(
-            initial_port=kwargs["port"],
+        register_filters(platform=platform, options=options)
+        options["port"] = find_serial_port(
+            initial_port=options["port"],
             board_config=platform.board_config(project_options.get("board"))
             if platform and project_options.get("board")
             else None,
             upload_protocol=project_options.get("upload_protocol"),
         )
 
-    # override system argv with patched options
-    sys.argv = ["monitor"] + project_options_to_monitor_argv(
-        kwargs,
-        project_options,
-        ignore=("port", "baud", "rts", "dtr", "environment", "project_dir"),
-    )
+    if options["menu_char"] == options["exit_char"]:
+        raise exception.UserSideException(
+            "--exit-char can not be the same as --menu-char"
+        )
 
-    if not kwargs["quiet"]:
-        click.echo(
-            "--- Available filters and text transformations: %s"
-            % ", ".join(sorted(miniterm.TRANSFORMATIONS.keys()))
-        )
-        click.echo("--- More details at https://bit.ly/pio-monitor-filters")
-    try:
-        miniterm.main(
-            default_port=kwargs["port"],
-            default_baudrate=kwargs["baud"]
-            or ProjectOptions["env.monitor_speed"].default,
-            default_rts=kwargs["rts"],
-            default_dtr=kwargs["dtr"],
-        )
-    except Exception as e:
-        raise exception.MinitermException(e)
+    start_terminal(options)
 
 
 def get_project_options(environment=None):
@@ -148,37 +141,13 @@ def get_project_options(environment=None):
     return config.items(env=environment, as_dict=True)
 
 
-def apply_project_monitor_options(cli_options, project_options):
+def apply_project_monitor_options(initial_options, project_options):
     for k in ("port", "speed", "rts", "dtr"):
         k2 = "monitor_%s" % k
         if k == "speed":
             k = "baud"
-        if cli_options[k] is None and k2 in project_options:
-            cli_options[k] = project_options[k2]
+        if initial_options[k] is None and k2 in project_options:
+            initial_options[k] = project_options[k2]
             if k != "port":
-                cli_options[k] = int(cli_options[k])
-    return cli_options
-
-
-def project_options_to_monitor_argv(cli_options, project_options, ignore=None):
-    confmon_flags = project_options.get("monitor_flags", [])
-    result = confmon_flags[::]
-
-    for f in project_options.get("monitor_filters", []):
-        result.extend(["--filter", f])
-
-    for k, v in cli_options.items():
-        if v is None or (ignore and k in ignore):
-            continue
-        k = "--" + k.replace("_", "-")
-        if k in confmon_flags:
-            continue
-        if isinstance(v, bool):
-            if v:
-                result.append(k)
-        elif isinstance(v, tuple):
-            for i in v:
-                result.extend([k, i])
-        else:
-            result.extend([k, str(v)])
-    return result
+                initial_options[k] = int(initial_options[k])
+    return initial_options